+++ sdk/highlight-run/src/client/otel/index.ts
@@ -331,6 +331,8 @@ export const setupBrowserTracing = (
 			],
 		}),
 	})
+
+	return providers
 }
 
 class CustomBatchSpanProcessor extends BatchSpanProcessor {