+++ sdk/highlight-run/src/client/types/types.ts
@@ -123,6 +123,18 @@ export type CommonOptions = {
 	 */
 	sessionCookie?: true
 	/**
+	 * Specifies whether Highlight will record user session replays.
+	 * Unless you are using Highlight only for error monitoring, you do not want to set this to true.
+	 * @default false
+	 */
+	disableSessionRecording?: boolean
+	/**
+	 * Specifies if Highlight should not automatically initialize when the class is created.
+	 * This should be used with `.start()` and `.stop()` if you want to control when Highlight records.
+	 * @default false
+	 */
+	manualStart?: boolean
+	/**
 	 * Key to use to combine sessions. If a sessionKey is provided, the session will be
 	 * combined with the previous session with the same key. Otherwise, the current session
 	 * logic will be used to generate a sessionSecureID.