+++ sdk/highlight-run/src/sdk/LDRecord.ts
@@ -11,11 +11,13 @@ class _LDRecord extends BufferedClass<Record> implements Record {
 	}
 
 	async start(options?: StartOptions) {
-		return this._bufferCall('start', [options])
+		// avoid buffering the start call
+		return this._sdk.start(options)
 	}
 
-	stop(options?: StartOptions) {
-		return this._bufferCall('stop', [options])
+	stop() {
+		// avoid buffering the stop call
+		return this._sdk.stop()
 	}
 
 	getRecordingState() {