+++ sdk/@launchdarkly/observability-node/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @launchdarkly/observability-node
 
+## 0.2.1
+
+### Patch Changes
+
+- b7145b5: Export options for node observability plugin.
+
+## 0.2.0
+
+### Minor Changes
+
+- 011032f: Change default configuration for instrumentation and expose more fine grained options.
+
 ## 0.1.2
 
 ### Patch Changes