+++ sdk/highlight-apollo/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @highlight-run/apollo
 
+## 3.4.45
+
+### Patch Changes
+
+- @highlight-run/node@3.12.18
+
+## 3.4.44
+
+### Patch Changes
+
+- @highlight-run/node@3.12.17
+
+## 3.4.43
+
+### Patch Changes
+
+- @highlight-run/node@3.12.16
+
 ## 3.4.42
 
 ### Patch Changes