+++ sdk/highlight-hono/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @highlight-run/hono
 
+## 1.0.21
+
+### Patch Changes
+
+- @highlight-run/node@3.12.18
+
+## 1.0.20
+
+### Patch Changes
+
+- @highlight-run/node@3.12.17
+
+## 1.0.19
+
+### Patch Changes
+
+- @highlight-run/node@3.12.16
+
 ## 1.0.18
 
 ### Patch Changes