+++ sdk/highlight-remix/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @highlight-run/remix
 
+## 2.0.117
+
+### Patch Changes
+
+- Updated dependencies [54aab8b]
+    - highlight.run@9.18.22
+    - @highlight-run/node@3.12.18
+    - @highlight-run/react@19.0.22
+
+## 2.0.116
+
+### Patch Changes
+
+- Updated dependencies [011032f]
+    - highlight.run@9.18.21
+    - @highlight-run/node@3.12.17
+    - @highlight-run/react@19.0.21
+
+## 2.0.115
+
+### Patch Changes
+
+- Updated dependencies [a151858]
+    - highlight.run@9.18.20
+    - @highlight-run/node@3.12.16
+    - @highlight-run/react@19.0.20
+
 ## 2.0.114
 
 ### Patch Changes